client/requirements.txt
cffi==1.15.1
cryptography==38.0.1
MouseInfo==0.1.3
Pillow==9.2.0
psutil==5.9.2
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python3-xlib==0.15
pytweening==1.0.4
